backend/pyproject.toml
[project]
name = "codescope-backend"
version = "0.1.0"
description = "CodeScope backend services"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "sqlalchemy>=2.0.40,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "alembic>=1.16.0,<2.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "pyjwt>=2.10.0,<3.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "bcrypt>=4.0.1,<4.1.0",
  "email-validator>=2.2.0,<3.0.0",
  "python-multipart>=0.0.20,<0.1.0",
  "celery>=5.4.0,<6.0.0",
  "redis>=5.2.0,<6.0.0",
  "neo4j>=5.26.0,<6.0.0",
  "minio>=7.2.8,<8.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
  "httpx>=0.28.0,<1.0.0",
  "ruff>=0.15.4",
]

[project.scripts]
codescope-worker = "app.worker.runner:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
